+++ pycontrails/models/cocip/cocip_params.py
@@ -152,8 +152,7 @@ class CocipParams(AdvectionBuffers):
     #: :attr:`CocipGridParams.verbose_outputs_evolution`.
     verbose_outputs: bool = False
 
-    #: Add effective radius to contrail.
-    #: It is not necessary for calculation.
+    #: Add effective radius to contrail output.
     output_effective_radius: bool = False
 
     #: Add additional metric of ATR20 and global yearly mean RF to model output.